# TSDB: Optimization: Merge postings using concrete type