Make `mergedPostings` generic, so we can call it with more specific types which is more efficient than making everything go through the `Postings` interface.

Move merge of head postings into index; remove the `Get` member which was not used for anything else, and fix up
tests.

~Unexport `PostingsFromDecbuf`; it looks like it should not have been exported.~

Preparatory work:
* Extend BenchmarkMemPostings_PostingsForLabelMatching to check merge speed.
* TSDB BenchmarkMerge: run fewer sizes.

Benchmark results:
```
goos: linux
goarch: amd64
pkg: github.com/prometheus/prometheus/tsdb/index
cpu: Intel(R) Core(TM) i7-14700K
                                                                   │ before.txt  │              after.txt              │
                                                                   │   sec/op    │    sec/op     vs base               │
Intersect/LongPostings1-28                                           10.70µ ± 0%    10.75µ ± 1%   +0.46% (p=0.026 n=6)
Intersect/LongPostings2-28                                           42.63m ± 2%    41.13m ± 1%   -3.53% (p=0.002 n=6)
Intersect/ManyPostings-28                                            111.9m ± 2%    111.0m ± 3%        ~ (p=0.699 n=6)
Merge/1-28                                                           193.5n ± 0%    177.8n ± 1%   -8.09% (p=0.002 n=6)
Merge/10-28                                                          9.947µ ± 0%   10.327µ ± 0%   +3.82% (p=0.002 n=6)
Merge/10000-28                                                       17.16m ± 2%    15.29m ± 3%  -10.92% (p=0.002 n=6)
Merge/100000-28                                                      241.8m ± 8%    199.3m ± 7%  -17.57% (p=0.002 n=6)
MemPostings_PostingsForLabelMatching/labels=1000/matcher=fast-28     27.89µ ± 0%    27.08µ ± 2%   -2.91% (p=0.002 n=6)
MemPostings_PostingsForLabelMatching/labels=1000/matcher=slow-28     4.894µ ± 4%    4.675µ ± 1%   -4.48% (p=0.002 n=6)
MemPostings_PostingsForLabelMatching/labels=1000/matcher=all-28      176.7µ ± 2%    159.1µ ± 0%   -9.95% (p=0.002 n=6)
MemPostings_PostingsForLabelMatching/labels=10000/matcher=fast-28    282.2µ ± 1%    282.0µ ± 1%        ~ (p=0.818 n=6)
MemPostings_PostingsForLabelMatching/labels=10000/matcher=slow-28    185.7m ± 2%    188.0m ± 3%        ~ (p=0.394 n=6)
MemPostings_PostingsForLabelMatching/labels=10000/matcher=all-28     2.055m ± 7%    1.804m ± 2%  -12.25% (p=0.002 n=6)
MemPostings_PostingsForLabelMatching/labels=100000/matcher=fast-28   2.902m ± 1%    2.842m ± 1%   -2.07% (p=0.004 n=6)
MemPostings_PostingsForLabelMatching/labels=100000/matcher=slow-28    2.371 ± 3%     2.331 ± 2%        ~ (p=0.240 n=6)
MemPostings_PostingsForLabelMatching/labels=100000/matcher=all-28    24.17m ± 1%    21.35m ± 1%  -11.67% (p=0.002 n=6)
PostingStatsMaxHep-28                                                54.28m ± 0%    34.04m ± 0%  -37.28% (p=0.002 n=6)
geomean                                                              1.715m         1.586m        -7.49%

                                                                   │    before.txt    │               after.txt                │
                                                                   │       B/op       │     B/op       vs base                 │
Intersect/LongPostings1-28                                              224.0 ±  0%       224.0 ±  0%        ~ (p=1.000 n=6) ¹
Intersect/LongPostings2-28                                              224.0 ±  0%       224.0 ±  0%        ~ (p=1.000 n=6) ¹
Intersect/ManyPostings-28                                               32.00 ±  0%       32.00 ±  0%        ~ (p=1.000 n=6) ¹
Merge/1-28                                                              0.000 ±  0%       0.000 ±  0%        ~ (p=1.000 n=6) ¹
Merge/10-28                                                             784.0 ±  0%       560.0 ±  0%  -28.57% (p=0.002 n=6)
Merge/10000-28                                                        632.1Ki ±  0%     472.1Ki ±  0%  -25.31% (p=0.002 n=6)
Merge/100000-28                                                       6.109Mi ±  0%     4.578Mi ±  0%  -25.06% (p=0.002 n=6)
MemPostings_PostingsForLabelMatching/labels=1000/matcher=fast-28      16.33Ki ±  0%     16.28Ki ±  0%   -0.29% (p=0.002 n=6)
MemPostings_PostingsForLabelMatching/labels=1000/matcher=slow-28      16.02Ki ±  0%     16.02Ki ±  0%        ~ (p=1.000 n=6) ¹
MemPostings_PostingsForLabelMatching/labels=1000/matcher=all-28       128.1Ki ±  0%     104.1Ki ±  0%  -18.74% (p=0.002 n=6)
MemPostings_PostingsForLabelMatching/labels=10000/matcher=fast-28     160.4Ki ±  0%     160.3Ki ±  0%   -0.03% (p=0.002 n=6)
MemPostings_PostingsForLabelMatching/labels=10000/matcher=slow-28     356.2Ki ± 55%     355.9Ki ± 55%        ~ (p=0.143 n=6)
MemPostings_PostingsForLabelMatching/labels=10000/matcher=all-28      1.242Mi ±  0%     1.008Mi ±  0%  -18.87% (p=0.002 n=6)
MemPostings_PostingsForLabelMatching/labels=100000/matcher=fast-28    1.532Mi ±  0%     1.532Mi ±  0%   -0.00% (p=0.002 n=6)
MemPostings_PostingsForLabelMatching/labels=100000/matcher=slow-28    2.483Mi ± 23%     2.971Mi ±  0%        ~ (p=1.000 n=6)
MemPostings_PostingsForLabelMatching/labels=100000/matcher=all-28    12.227Mi ±  0%     9.930Mi ±  0%  -18.79% (p=0.002 n=6)
PostingStatsMaxHep-28                                                   240.0 ±  0%       240.0 ±  0%        ~ (p=1.000 n=6) ¹
geomean                                                                             ²                   -7.72%               ²
¹ all samples are equal
² summaries must be >0 to compute geomean

                                                                   │   before.txt    │               after.txt                │
                                                                   │    allocs/op    │   allocs/op     vs base                │
Intersect/LongPostings1-28                                            6.000 ±   0%       6.000 ±   0%       ~ (p=1.000 n=6) ¹
Intersect/LongPostings2-28                                            6.000 ±   0%       6.000 ±   0%       ~ (p=1.000 n=6) ¹
Intersect/ManyPostings-28                                             1.000 ±   0%       1.000 ±   0%       ~ (p=1.000 n=6) ¹
Merge/1-28                                                            0.000 ±   0%       0.000 ±   0%       ~ (p=1.000 n=6) ¹
Merge/10-28                                                           3.000 ±   0%       3.000 ±   0%       ~ (p=1.000 n=6) ¹
Merge/10000-28                                                        3.000 ±   0%       3.000 ±   0%       ~ (p=1.000 n=6) ¹
Merge/100000-28                                                       3.000 ±   0%       3.000 ±   0%       ~ (p=1.000 n=6) ¹
MemPostings_PostingsForLabelMatching/labels=1000/matcher=fast-28      6.000 ±   0%       6.000 ±   0%       ~ (p=1.000 n=6) ¹
MemPostings_PostingsForLabelMatching/labels=1000/matcher=slow-28      2.000 ±   0%       2.000 ±   0%       ~ (p=1.000 n=6) ¹
MemPostings_PostingsForLabelMatching/labels=1000/matcher=all-28       6.000 ±   0%       6.000 ±   0%       ~ (p=1.000 n=6) ¹
MemPostings_PostingsForLabelMatching/labels=10000/matcher=fast-28     6.000 ±   0%       6.000 ±   0%       ~ (p=1.000 n=6) ¹
MemPostings_PostingsForLabelMatching/labels=10000/matcher=slow-28    2.042k ± 100%      2.032k ± 100%       ~ (p=0.147 n=6)
MemPostings_PostingsForLabelMatching/labels=10000/matcher=all-28      6.000 ±   0%       6.000 ±   0%       ~ (p=1.000 n=6) ¹
MemPostings_PostingsForLabelMatching/labels=100000/matcher=fast-28    6.000 ±   0%       6.000 ±   0%       ~ (p=1.000 n=6) ¹
MemPostings_PostingsForLabelMatching/labels=100000/matcher=slow-28   6.129k ± 100%     12.186k ±   0%       ~ (p=1.000 n=6)
MemPostings_PostingsForLabelMatching/labels=100000/matcher=all-28     6.000 ±   0%       6.000 ±   0%       ~ (p=1.000 n=6) ¹
PostingStatsMaxHep-28                                                 1.000 ±   0%       1.000 ±   0%       ~ (p=1.000 n=6) ¹
geomean                                                                            ²                   +4.10%               ²
```