@aknuds1 2024-09-21
> Intersect/LongPostings1 slowdown seems consistent, although it was not changed in this PR. I will investigate further.

I also noticed this regression. Could you also rebase and fix the conflict(s)?