@bboreham 2024-12-20
I have rebased and updated the benchmarks.  The regression seems to have gone away.

I can no longer apply the same optimisation to `Reader` methods, because the indirection added in #13567 means we don't know what type each postings is there.