@bboreham 2025-01-02
I ran the TSDB `PostingsForMatchers` benchmarks, and see what looks like a large percentage slow-down in many cases.
However only the ones that take a couple of hundred nanoseconds slow down; all the ones that take longer speed up.
The reason is that the `Get` path did fewer allocations than the `Postings` path does now.

I could put back that case - check `len(values) == 1` in `Postings` - but I don't think it's worth it for 50ns per query.

```
goos: linux
goarch: amd64
pkg: github.com/prometheus/prometheus/tsdb
cpu: Intel(R) Core(TM) i7-14700K
                                                                     │  before2.txt  │             after2.txt              │
                                                                     │    sec/op     │    sec/op     vs base               │
Querier/Head/PostingsForMatchers/n="1"-28                               203.8n ±  5%   229.6n ±  1%  +12.66% (p=0.002 n=6)
Querier/Head/PostingsForMatchers/n="X"-28                               189.8n ±  1%   244.9n ±  2%  +29.06% (p=0.002 n=6)
Querier/Head/PostingsForMatchers/n="1",j="foo"-28                       322.0n ±  1%   385.6n ±  1%  +19.74% (p=0.002 n=6)
Querier/Head/PostingsForMatchers/n="X",j="foo"-28                       220.7n ±  7%   271.6n ±  0%  +23.09% (p=0.002 n=6)
Querier/Head/PostingsForMatchers/j="foo",n="1"-28                       352.3n ±  1%   387.7n ±  1%  +10.03% (p=0.002 n=6)
Querier/Head/PostingsForMatchers/n="1",j!="foo"-28                      342.3n ±  1%   378.8n ±  1%  +10.66% (p=0.002 n=6)
Querier/Head/PostingsForMatchers/n="1",i!="2"-28                        346.6n ±  1%   383.3n ±  1%  +10.60% (p=0.002 n=6)
Querier/Head/PostingsForMatchers/n="X",j!="foo"-28                      225.1n ±  1%   262.7n ±  0%  +16.73% (p=0.002 n=6)
Querier/Head/PostingsForMatchers/i=~"1[0-9]",j=~"foo|bar"-28            679.6n ±  1%   463.0n ±  5%  -31.87% (p=0.002 n=6)
Querier/Head/PostingsForMatchers/j=~"foo|bar"-28                        392.2n ±  2%   357.4n ±  0%   -8.86% (p=0.002 n=6)
Querier/Head/PostingsForMatchers/j=~"XXX|YYY"-28                        314.3n ±  0%   283.0n ±  1%   -9.96% (p=0.002 n=6)
Querier/Head/PostingsForMatchers/j=~"X.+"-28                            247.3n ±  1%   242.1n ±  0%   -2.08% (p=0.002 n=6)
Querier/Head/PostingsForMatchers/i=~"(1|2|3|4|5|6|20|55)"-28            590.2n ±  1%   395.1n ±  2%  -33.07% (p=0.002 n=6)
Querier/Head/PostingsForMatchers/i!~"(1|2|3|4|5|6|20|55)"-28            632.9n ±  1%   474.2n ±  2%  -25.06% (p=0.002 n=6)
Querier/Head/PostingsForMatchers/i=~"X|Y|Z"-28                          367.5n ±  1%   312.0n ±  0%  -15.11% (p=0.002 n=6)
Querier/Head/PostingsForMatchers/i!~"X|Y|Z"-28                          423.8n ±  0%   391.2n ±  1%   -7.68% (p=0.002 n=6)
Querier/Head/PostingsForMatchers/i=~".*"-28                             186.9n ±  1%   205.1n ±  0%   +9.74% (p=0.002 n=6)
Querier/Head/PostingsForMatchers/i=~"1.*"-28                            1.671m ±  6%   1.588m ±  5%   -4.98% (p=0.026 n=6)
Querier/Head/PostingsForMatchers/i=~".*1"-28                            1.044m ± 15%   1.084m ±  3%        ~ (p=0.310 n=6)
Querier/Head/PostingsForMatchers/i=~".+"-28                             3.966m ± 11%   3.302m ±  8%  -16.73% (p=0.004 n=6)
Querier/Head/PostingsForMatchers/i=~".+",j=~"X.+"-28                    3.865m ±  5%   3.410m ±  5%  -11.76% (p=0.002 n=6)
Querier/Head/PostingsForMatchers/i=~""-28                               3.986m ±  5%   3.383m ±  3%  -15.13% (p=0.002 n=6)
Querier/Head/PostingsForMatchers/i!=""-28                               3.824m ±  8%   3.386m ± 10%  -11.45% (p=0.002 n=6)
Querier/Head/PostingsForMatchers/n="1",i=~".*",j="foo"-28               375.6n ±  4%   413.1n ±  3%   +9.96% (p=0.002 n=6)
Querier/Head/PostingsForMatchers/n="X",i=~".*",j="foo"-28               256.5n ±  1%   295.2n ±  0%  +15.09% (p=0.002 n=6)
Querier/Head/PostingsForMatchers/n="1",i=~".*",i!="2",j="foo"-28        518.2n ±  2%   568.6n ±  1%   +9.73% (p=0.002 n=6)
Querier/Head/PostingsForMatchers/n="1",i!=""-28                         3.919m ± 10%   3.481m ±  7%  -11.16% (p=0.002 n=6)
Querier/Head/PostingsForMatchers/n="1",i!="",j="foo"-28                 3.905m ±  8%   3.377m ±  7%  -13.51% (p=0.002 n=6)
Querier/Head/PostingsForMatchers/n="1",i!="",j=~"X.+"-28                3.852m ±  7%   3.537m ±  7%   -8.18% (p=0.009 n=6)
Querier/Head/PostingsForMatchers/n="1",i!="",j=~"XXX|YYY"-28            3.927m ±  3%   3.812m ±  9%        ~ (p=0.093 n=6)
Querier/Head/PostingsForMatchers/n="1",i=~"X|Y|Z",j="foo"-28            528.6n ±  2%   489.4n ±  2%   -7.42% (p=0.002 n=6)
Querier/Head/PostingsForMatchers/n="1",i!~"X|Y|Z",j="foo"-28            630.9n ±  1%   601.0n ±  1%   -4.74% (p=0.002 n=6)
Querier/Head/PostingsForMatchers/n="1",i=~".+",j="foo"-28               3.965m ±  4%   3.457m ±  4%  -12.82% (p=0.002 n=6)
Querier/Head/PostingsForMatchers/n="1",i=~"1.+",j="foo"-28              1.629m ±  4%   1.623m ±  5%        ~ (p=0.699 n=6)
Querier/Head/PostingsForMatchers/n="1",i=~".*1.*",j="foo"-28            5.364m ±  5%   4.917m ±  4%   -8.33% (p=0.002 n=6)
Querier/Head/PostingsForMatchers/n="1",i=~".+",i!="2",j="foo"-28        3.927m ±  8%   3.446m ±  5%  -12.24% (p=0.002 n=6)
Querier/Head/PostingsForMatchers/n="1",i=~".+",i!~"2.*",j="foo"-28      5.784m ±  7%   5.493m ±  6%   -5.03% (p=0.026 n=6)
Querier/Head/PostingsForMatchers/n="1",i=~".+",i!~".*2.*",j="foo"-28   10.065m ± 12%   8.482m ±  2%  -15.73% (p=0.002 n=6)
Querier/Head/PostingsForMatchers/n="X",i=~".+",i!~".*2.*",j="foo"-28    306.2n ±  1%   346.0n ±  2%  +13.00% (p=0.002 n=6)
geomean                                                                 15.43µ         14.86µ         -3.70%
```